+++ examples/resources/identitycloud_secret/resource.tf
@@ -1,6 +1,7 @@
 resource "identitycloud_secret" "example" {
-  variable_id  = "esv-mysecret1"
-  description  = "My secret"
-  encoding     = "generic"
-  value_base64 = base64encode("secretvalue")
+  secret_id           = "esv-examplesecret"
+  description         = "my example secret"
+  encoding            = "generic"
+  use_in_placeholders = false
+  value_base64        = base64encode(var.example_secret_value)
 }