+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 * **New Resource:** `identitycloud_content_security_policy_enforced` ([#23](https://github.com/pingidentity/terraform-provider-pingfederate/pull/23))
 * **New Resource:** `identitycloud_content_security_policy_report_only` ([#23](https://github.com/pingidentity/terraform-provider-pingfederate/pull/23))
 * **New Resource:** `identitycloud_promotion_lock` ([#24](https://github.com/pingidentity/terraform-provider-pingfederate/pull/24))
+* **New Resource:** `identitycloud_secret_version` ([#32](https://github.com/pingidentity/terraform-provider-pingfederate/pull/32))
 
 ### Data sources
 * **New Data Source:** `identitycloud_promotion_lock` ([#24](https://github.com/pingidentity/terraform-provider-pingfederate/pull/24))
+* **New Data Source:** `identitycloud_secret_versions` ([#32](https://github.com/pingidentity/terraform-provider-pingfederate/pull/32))